JSON Formatting  for Date.

+++ src/main/java/com/globalsoftwaresupport/repository/UserRepository.java
@@ -23,7 +23,7 @@ public class UserRepository {
     }
 
     public void save(User user) {
-        user.setId(users.size());
+        user.setId((long)users.size());
         this.users.add(user);
     }
